-The Relation "obj" made it to point to propositions
-The Relation "doo" was used in some of the acts instead of the
relation"obj",as the relation "doo" points to acts.example of these acts
are doOne,doAll.

-The processReports() method in the node class(which was empty,has no
body) was commented because there
is an identical method that is in ActNode class but have different
return type,thus a contradiction.It
was also commented in the runner.

-The control acts were modified to fit the new implementation

-There is a method in network class buildActionNode(String
identifier,String type),It creates
an actionNode that with a string identfier for the node and another
string type that specfies
the type of the action node that will be created.

-In testing,there is a method commented testPlans that tests plans.It
was tested on a created class that was non-primitive,but the class was
deleted since it is only purpose was the testing,and thus the
method is commented

-In testing, SNIterate test fails because the guards are not believed,If
they are to be believed
It will keep looping and running (as the act should do).The act stops
when there is no guards
are believed.Thus one way to fulfill the test is to disBelieve all the
guards at the end of act method.